11/21/2024 Added final score and additional reset instructors upon game end. Final score color is conditional to the final score. Red, Yellow, Green.
Needs to be done:
Logic to make all pipes possible. Some are too far away.
First pipe is impossible 50% of the time because it is too low. Add logic to where if self.score == 0, the pipes have to be in a central position.

Ideas:
Main Menu
Difficulty modes
Sound
Highscore